GitLab and TravisCI can no longer build versions of this package
GitLab now is failing on PHP versions 7.0, 7.1, 7.3, and 7.4 due to an error installing the PHP ZIP extension.

TravisCI is failing on PHP versions 7.3 and 7.4 due to using XDebug version 3.0 and the constants not being compatible with the PHPUnit use for testing.

This is no fault of the package, but of the test runners.

Therefore, I'm removing the badges of the GitLab CI and TravisCI, and adding the badge for GitHub Actions.